requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.11.1
Flask==2.2.2
imgurpython==1.1.7
line_bot_sdk==1.14.0
matplotlib==3.5.3
pandas==1.5.1
pandas_datareader==0.10.0
pyimgur==0.6.0
python-dotenv==0.21.0
requests==2.28.1
transitions==0.6.9
yfinance==0.2.3

chardet==3.0.4
future==0.18.1
six==1.12.0
urllib3==1.25.6